General fixes and improving code readability this also adds a check for makepkg
This commit fixes major issues I stumbled upon while testing including but not limited to: 
line 7 $ for no reason
broken imput from user on lines 11 and 43

this also adds some small improvements to my script like "cd ~" at the start and at the end 

This commit implements a check for makepkg ( a script used in Arch-Linux in order to compile software ).
If makepkg isn't found, it will be installed.

All in all, this is a pretty big and important commit (for me).  I hope it stands the test of time.